+++ ORSProject-04/src/main/java/in/co/rays/proj4/model/MarksheetModel.java
@@ -146,14 +146,14 @@ public class MarksheetModel {
 		}
 	}
 
-	public void delete(int id) throws ApplicationException {
+	public void delete(long id) throws ApplicationException {
 		Connection conn = null;
 		try {
 			conn = JDBCDataSource.getConnection();
 			conn.setAutoCommit(false);
 
 			PreparedStatement pstmt = conn.prepareStatement("delete from st_marksheet where id=?");
-			pstmt.setInt(1, id);
+			pstmt.setLong(1, id);
 			int i = pstmt.executeUpdate();
 			System.out.println("Data Deleted => " + i);
