+++ ORSProject-04/src/main/java/in/co/rays/proj4/model/SubjectModel.java
@@ -138,14 +138,14 @@ public class SubjectModel {
 		}
 	}
 
-	public void delete(int id) throws ApplicationException {
+	public void delete(long id) throws ApplicationException {
 		Connection conn = null;
 		try {
 			conn = JDBCDataSource.getConnection();
 			conn.setAutoCommit(false);
 
 			PreparedStatement pstmt = conn.prepareStatement("delete from st_subject where id=?");
-			pstmt.setInt(1, id);
+			pstmt.setLong(1, id);
 			int i = pstmt.executeUpdate();
 			System.out.println("Data Deleted => " + i);
